h2: fix warning on window update check

+++ lib/http2/http2.c
@@ -813,7 +813,7 @@ lws_h2_parse_frame_header(struct lws *wsi)
 		h2n->swsi->h2.peer_tx_cr_est -= h2n->length;
 		lwsl_debug("   peer_tx_cr_est %d\n",
 			   h2n->swsi->h2.peer_tx_cr_est);
-		if (h2n->swsi->h2.peer_tx_cr_est < h2n->length + 265536) {
+		if (h2n->swsi->h2.peer_tx_cr_est < (int32_t)h2n->length + 265536) {
 			h2n->swsi->h2.peer_tx_cr_est += h2n->length + 265536;
 			pps = lws_h2_new_pps(LWS_H2_PPS_UPDATE_WINDOW);
 			if (!pps)
